wannabe cqrs
