+++ app/server/queries/petitionsStats.js
