+++ support/testing/index.js
@@ -1,3 +1,2 @@
 export * from './eventually';
 export * from './page';
-export * from './seeding';
